requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
python-dotenv